Preparations for optimizations.

+++ src/sig/Panel.java
@@ -9,19 +9,15 @@ import java.awt.Color;
 import java.awt.Image;
 import java.awt.image.MemoryImageSource;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.awt.image.ColorModel;
 import java.awt.GraphicsEnvironment;
 import java.awt.GraphicsConfiguration;
 import java.awt.Toolkit;
-import java.util.Queue;
 
 public class Panel extends JPanel implements Runnable {
     long startTime = System.nanoTime();
@@ -270,9 +266,9 @@ public class Panel extends JPanel implements Runnable {
                     }
                     SigRenderer.temp_request=SigRenderer.request;
                     DrawUtils.TexturedTriangle(p, 
-                        (int)tt.A.x,(int)tt.A.y,tt.T.u,tt.T.v,tt.T.w,
-                        (int)tt.B.x,(int)tt.B.y,tt.U.u,tt.U.v,tt.U.w,
-                        (int)tt.C.x,(int)tt.C.y,tt.V.u,tt.V.v,tt.V.w,
+                        tt.A.x,tt.A.y,tt.T.u,tt.T.v,tt.T.w,
+                        tt.B.x,tt.B.y,tt.U.u,tt.U.v,tt.U.w,
+                        tt.C.x,tt.C.y,tt.V.u,tt.V.v,tt.V.w,
                     tt.tex,(tt.col&0xFF0000)>>16,tt);
                 } else {
                     DrawUtils.FillTriangle(p,(int)tt.A.x,(int)tt.A.y,(int)tt.B.x,(int)tt.B.y,(int)tt.C.x,(int)tt.C.y,tt.getColor());
